# Optimize VideoCapture with cvtColor instead of sws_scale