@dkurt 2025-09-07
see https://github.com/opencv/opencv/pull/27755